Move "job" related code to "job" module
In the context of https://github.com/wbsoft/frescobaldi/projects/4
This is the first step of encapsulating all job related code into a module,
which will make extending the functionality substantially easier:
https://github.com/wbsoft/frescobaldi/projects/4#card-12637526

The next steps in this specific sub-topic of the project will be
a)
making clearer that a "job" can be anything, not only a LilyPond job
(admittedly I'm right now not sure if that's not the case anyway)
b)
implementing a multithreaded job queue (which I actually have done
in a private fork for a project and will make generically available)
