+++ frescobaldi_app/musicview/__init__.py
@@ -51,7 +51,7 @@ import qutil
 import panel
 import listmodel
 import gadgets.drag
-import jobattributes
+from job import attributes as jobattributes
 
 from . import documents
